+++ syncpipe/sync_pipe.go
@@ -1,4 +1,4 @@
-package namespaces
+package syncpipe
 
 import (
 	"encoding/json"
