+++ syncpipe/sync_pipe_linux.go
@@ -1,4 +1,4 @@
-package namespaces
+package syncpipe
 
 import (
 	"os"
