+++ syncpipe/sync_pipe_test.go
@@ -1,4 +1,4 @@
-package namespaces
+package syncpipe
 
 import (
 	"fmt"
